Data_Generation_Preparation/requirements.txt
ackermann-msgs==2.0.2
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.2.4
addict==2.4.0
ament-cmake-test==1.3.10
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
angles==1.15.0
asttokens==2.4.1
attrs==23.2.0
autocommand==2.2.2
backports.tarfile==1.2.0
blinker==1.8.2
-e git+https://github.com/thodan/bop_toolkit.git@e19fde2c5a8f99f4629bb7e6ce69c7339cc2dc70#egg=bop_toolkit_lib
braceexpand==0.1.7
builtin-interfaces==1.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
composition-interfaces==1.2.1
ConfigArgParse==1.7
contourpy==1.2.1
control-msgs==4.6.0
controller-manager==2.42.0
controller-manager-msgs==2.42.0
cv-bridge==3.2.1
cycler==0.12.1
Cython==3.0.10
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.2.4
domain-coordinator==0.10.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.2
examples-rclpy-minimal-action-client==0.15.2
examples-rclpy-minimal-action-server==0.15.2
examples-rclpy-minimal-client==0.15.2
examples-rclpy-minimal-publisher==0.15.2
examples-rclpy-minimal-service==0.15.2
examples-rclpy-minimal-subscriber==0.15.2
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
Flask==3.0.3
fonttools==4.53.1
freetype-py==2.4.0
gazebo-msgs==3.7.0
generate-parameter-library-py==0.3.8
geometry-msgs==4.2.4
glumpy==1.2.1
hsluv==5.0.4
idna==3.7
image-geometry==3.2.1
imageio==2.34.2
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
interactive-markers==2.3.2
ipython==8.26.0
ipywidgets==8.1.3
itsdangerous==2.2.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
laser-geometry==2.4.0
launch==1.0.6
launch-param-builder==0.1.1
launch-ros==0.19.7
launch-testing==1.0.6
launch-testing-ros==0.19.7
launch-xml==1.0.6
launch-yaml==1.0.6
lifecycle-msgs==1.2.1
logging-demo==0.20.5
map-msgs==2.1.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
message-filters==4.3.4
more-itertools==10.3.0
moveit-configs-utils==2.5.5
moveit-msgs==2.2.1
nav-msgs==4.2.4
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.0.1
object-recognition-msgs==2.0.0
octomap-msgs==2.0.0
open3d==0.18.0
opencv-python==4.10.0.84
ordered-set==4.1.0
osrf-pycommon==2.0.2
packaging==24.1
pandas==2.2.2
parso==0.8.4
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools @ file:///home/utsav/IProject/bop_toolkit/cocoapi/PythonAPI
Pygments==2.18.0
PyOpenGL==3.1.0
pyparsing==3.1.2
pypng==0.20220715.0
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-qt-binding==1.1.2
pytz==2024.1
PyYAML==6.0.1
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.14
rcutils==5.1.6
referencing==0.35.1
requests==2.32.3
resource-retriever==3.1.2
retrying==1.3.4
rmw-dds-common==1.6.0
ros2-controllers-test-nodes==2.37.0
ros2action==0.18.11
ros2bag==0.15.12
ros2cli==0.18.11
ros2component==0.18.11
ros2controlcli==2.42.0
ros2doctor==0.18.11
ros2interface==0.18.11
ros2launch==0.19.7
ros2lifecycle==0.18.11
ros2multicast==0.18.11
ros2node==0.18.11
ros2param==0.18.11
ros2pkg==0.18.11
ros2run==0.18.11
ros2service==0.18.11
ros2topic==0.18.11
rosbag2-interfaces==0.15.12
rosbag2-py==0.15.12
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpds-py==0.19.1
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-joint-trajectory-controller==2.37.0
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scikit-learn==1.5.1
scipy==1.14.0
scripts==3.7.0
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
shape-msgs==4.2.4
six==1.16.0
srdfdom==2.0.4
sros2==0.10.5
stack-data==0.6.3
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
teleop-twist-keyboard==2.4.0
tenacity==9.0.0
tf2-geometry-msgs==0.25.7
tf2-kdl==0.25.7
tf2-msgs==0.25.7
tf2-py==0.25.7
tf2-ros-py==0.25.7
tf2-tools==0.25.7
threadpoolctl==3.5.0
tomli==2.0.1
topic-monitor==0.20.5
tqdm==4.66.4
traitlets==5.14.3
trajectory-msgs==4.2.4
triangle==20230923
turtle-tf2-py==0.3.7
turtlesim==1.4.2
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
unique-identifier-msgs==2.2.1
ur-dashboard-msgs==2.2.15
ur-moveit-config==2.2.15
ur-msgs==2.0.0
ur-robot-driver==2.2.15
urdfdom-py==1.2.1
urllib3==2.2.2
vispy==0.14.3
visualization-msgs==4.2.4
wcwidth==0.2.13
webdataset==0.2.86
Werkzeug==3.0.3
widgetsnbextension==4.0.11
xacro==2.0.8
zipp==3.19.2
